refactor(illustration): rename scss files to use .module.scss
