+++ draft-packages/illustration/KaizenDraft/Illustration/Base.module.scss
@@ -1,6 +1,6 @@
 .wrapper {
   width: 100%;
-  margin: 0; // Override global Material UI styles
+  margin: 0;
 }
 
 .aspectRatioWrapper {
