+++ draft-packages/illustration/KaizenDraft/Illustration/Players/VideoPlayer.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useEffect } from "react"
 import { assetUrl } from "@kaizen/hosted-assets"
-import styles from "../style.module.scss"
 import { canPlayWebm } from "../utils"
+import styles from "../Base.module.scss"
 
 export type VideoPlayerProps = {
   /**
