+++ draft-packages/illustration/package.json
@@ -33,6 +33,7 @@
   "license": "MIT",
   "dependencies": {
     "@kaizen/hosted-assets": "^2.0.0",
+    "classnames": "^2.3.1",
     "jszip": "^3.7.1",
     "lottie-web": "^5.7.14"
   },
